started work on solver program. created a main and a function that checks the given arguments for errors. also modified ft_atoi to return a long long instead of an int to help check if an arguments is outside the int range of values

+++ libft/ft_atoi.c
@@ -20,7 +20,7 @@ static int	is_whitespace(char c)
 		return (0);
 }
 
-int			ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	int					i;
 	unsigned long long	res;
